requirements.txt
streamlit==1.28.1
Pillow==9.5.0